+++ demo/views/actions/actions.js
@@ -5,6 +5,7 @@ import SplitButtonDemo from './split-button-demo';
 import MultiActionButtomDemo from './multi-action-button-demo';
 import MultiStepWizardDemo from './multi-step-wizard-demo';
 import ButtonToggleDemo from './button-toggle-demo';
+import CreateDemo from './create-demo';
 import MenuDemo from './menu-demo';
 import AnimatedMenuButtonDemo from './animated-menu-button-demo';
 
@@ -22,6 +23,7 @@ class Actions extends React.Component {
         <MultiStepWizardDemo />
         <SplitButtonDemo />
         <ButtonToggleDemo />
+        <CreateDemo />
         <MenuDemo />
         <AnimatedMenuButtonDemo />
       </div>